+++ src/main/java/com/mysql/example/demo/controllers/nba/GamesByDateController.java
@@ -1,4 +1,4 @@
-package com.mysql.example.demo;
+package com.mysql.example.demo.controllers.nba;
 
 import java.util.List;
 
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.mysql.example.demo.IGamesByDateClientService;
 import com.mysql.example.demo.responses.GamesByDate;
 import com.mysql.example.demo.responses.mobile.GamesByDateMobileResponse;
 import com.mysql.example.demo.services.interfaces.IBackendRequestService;
@@ -17,6 +18,7 @@ import com.mysql.example.demo.services.interfaces.IBackendRequestService;
 public class GamesByDateController {
     private final IBackendRequestService requests;
     private final IGamesByDateClientService gamesService;
+    
     @Autowired
     public GamesByDateController(IBackendRequestService requests, IGamesByDateClientService gamesService) {
         this.requests = requests;
